認証と一対多をタスク管理アプリケーションに追加

+++ app/Http/Controllers/TasksController.php
@@ -13,26 +13,39 @@ class TasksController extends Controller
      */
     public function index()
     {
-        //タスク一覧を取得
-        $tasks = Task::all();         // 追加
+        $data = [];
+        if (\Auth::check()) {
+            
+            $user = \Auth::user();
+            $tasks = Task::where('user_id',$user->id)->get();
+            
+            $data = [
+                    'user' => $user,
+                    'tasks' => $tasks,
+                ];
 
-        // タスク一覧ビューでそれを表示
-        return view('tasks.index', [     // 追加
-            'tasks' => $tasks,        // 追加
-        ]);                                 // 追加
+            return view('tasks.index', [     // 追加
+                'tasks' => $tasks,        // 追加
+        ], $data);
+        }
+        return redirect('/');
+                                  
     }
 
     /**
      * Show the form for creating a new resource.
      */
     public function create()
-    {
-         $task = new Task;
+    {   
+        if (\Auth::check()) {
+            $task = new Task;
 
-        // タスク作成ビューを表示
-        return view('tasks.create', [
-            'task' => $task,
-        ]);
+            // タスク作成ビューを表示
+            return view('tasks.create', [
+                'task' => $task,
+            ]);
+       }
+       return view('dashboard');
     }
 
     /**
@@ -40,27 +53,36 @@ class TasksController extends Controller
      */
     public function store(Request $request)
     {
-         $request->validate([
-            'content' => 'required|max:255',
-            'status' => 'required|max:10',
-        ]);
-        $task = new Task;
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
-
-         return redirect('/');
+        if (\Auth::check()) {
+            $request->validate([
+                'content' => 'required|max:255',
+                'status' => 'required|max:10',
+            ]);
+            $user = \Auth::user();
+            $userId = $user->id;
+            $task = new Task([
+            'content' => $request->content,
+            'status' => $request->status,
+            'user_id' => $userId,
+            ]);
+            $task->save();
+            return redirect('/top');
+         }
+        return redirect('/');
     }
 
     /**
      * Display the specified resource.
      */
     public function show(string $id)
-    {
-        $task = Task::findOrFail($id);
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+    {   
+        if (\Auth::check()) {
+            $task = Task::findOrFail($id);
+            return view('tasks.show', [
+                'task' => $task,
+            ]);
+        } 
+       return redirect('/');
     }
 
     /**
@@ -68,34 +90,61 @@ class TasksController extends Controller
      */
     public function edit($id)
     {
-        $task = Task::findOrFail($id);
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
+        if (\Auth::check()) {
+            $task = Task::findOrFail($id);
+            return view('tasks.edit', [
+                'task' => $task,
+            ]);
+       }
+       return redirect('/');
     }
 
     /**
      * Update the specified resource in storage.
      */
     public function update(Request $request, $id)
-    {
-         $request->validate([
-            'content' => 'required|max:255',
-            'status' => 'required|max:10',
-        ]);
-        $task = Task::findOrFail($id);
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
+    {   
+        if (\Auth::check()) {
+            $request->validate([
+                'content' => 'required|max:255',
+                'status' => 'required|max:10',
+            ]);
+            $task = Task::findOrFail($id);
+            $user = \Auth::user();
+            $userId = $user->id;
+            if(!($task->user_id == $userId)){
+                return redirect('/top');
+            }
+            $task->content = $request->content;
+            $task->status = $request->status;
+            $task->save();
+            return redirect('/top');
+        }
         return redirect('/');
     }
     /**
      * Remove the specified resource from storage.
      */
     public function destroy($id)
-    {
-        $task = Task::findOrFail($id);
-        $task->delete();
+    { 
+        if (\Auth::check()) {
+            $task = Task::findOrFail($id);
+            $user = \Auth::user();
+            $userId = $user->id;
+            if(!($task->user_id == $userId)){
+                return redirect('/top');
+            }
+            $task->delete();
+            return redirect('/top');
+        }
         return redirect('/');
     }
+
+    public function checkAuth($id)
+    { 
+        if (\Auth::check()) {
+            return redirect('/top');
+        }
+    }
+
 }
